requirements.txt
streamlit
pandas
pymongo == 3.12
confluent-kafka
plotly
python-dotenv
dnspython
